embers v0.1.10 Embers.Notifications

Notifications context

A notification is an event that is sent to the user and stored, such as new comments, follows or mentions.

A notification has the following members:

:type representing the notification type, meant to be used for correct frontend representation

:user_id from the emitter

:source_id that points to the source, if any(eg: a comment)

:read status, being true or false

:text if needed

timestamps inserted_at and updated_at

Notification types are validated against a @valid_types list in Embers.Feed.Notification module(wich is where the Notification schema is defined)

Functions

create_notification(attrs)

create_notification(map()) ::
  {:ok, Embers.Notifications.Notification.t()} | {:error, Ecto.Changeset.t()}

Creates a notification.

Examples

iex> create_notification(%{"field" => value})
{:ok, %Notification{}}

iex> create_notification(%{"field" => bad_value})
{:error, %Ecto.Changeset{}}

delete_notification(id)

delete_notification(integer()) :: Embers.Notifications.Notification.t() | nil

Deletes a notification by it’s id and returns Notification or nil

Examples

iex> delete_notification(123)
%Notification{}

iex> delete_notification(456)
nil

get!(id)

get!(integer()) :: Embers.Notifications.Notification.t()

Returns a Notification by it’s :id or raises if nor found

Examples

iex> get!(123)
%Notification{}

iex> get!(456)
** (Ecto.NoResultsError)

get(id)

get(integer()) :: Embers.Notifications.Notification.t() | nil

Returns a Notification by it’s :id or nil if not found

Examples

iex> get(123)
%Notification{}

iex> get(456)
nil