embers v0.1.10 Embers.Paginator

Functions

paginate(queryable, opts \\ %{})

Fetches all the results matching the query between the cursors

It relies on the id column for the cursor

Options

after - Fetch the records after this id

before - Fetch the records before this id

limit - Limits the number of records returned per page. Note that this number will be capped by :max_limit. Defaults to 50

max_limit - Sets a maximum cap for :limit. This option can be useful when :limit is set dynamically (e.g from a URL param set by a user) but you still want to enfore a maximum. Defaults to 500.

Usage

query
|> Paginator.paginate(opts)