embers v0.1.18 EmbersWeb.Email

A module for sending emails to the user. This module provides functions to be used with the Phauxth authentication library when confirming users or handling password resets. It uses Bamboo, with the Mandrill adapter, to email users. For tests, it uses a test adapter, which is configured in the config/test.exs file. If you want to use a different email adapter, or another email library, read the instructions below.

Bamboo with a different adapter

Bamboo has adapters for Mailgun, Mailjet, Mandrill, Sendgrid, SMTP, SparkPost, PostageApp, Postmark and Sendcloud. There is also a LocalAdapter, which is great for local development. See Bamboo for more information.

Other email library

If you do not want to use Bamboo, follow the instructions below:

Edit this file, using the email library of your choice

Remove the lib/forks_the_egg_sample/mailer.ex file

Remove the Bamboo entries in the config/config.exs and config/test.exs files

Remove bamboo from the deps section in the mix.exs file

Functions

confirm_request(address, key)

An email with a confirmation link in it.

confirm_success(address)

An email acknowledging that the account has been successfully confirmed.

reset_request(address, key)

An email with a link to reset the password.

reset_success(address)

An email acknowledging that the password has been successfully reset.