embers v0.1.18 EmbersWeb.Router.Helpers

Module with named helpers generated from EmbersWeb.Router.

Functions

account_path(conn_or_endpoint, action)

account_path(conn_or_endpoint, action, params)

account_url(conn_or_endpoint, action)

account_url(conn_or_endpoint, action, params)

block_path(conn_or_endpoint, action)

block_path(conn_or_endpoint, action, params)

block_path(conn_or_endpoint, action, id, params)

block_url(conn_or_endpoint, action)

block_url(conn_or_endpoint, action, params)

block_url(conn_or_endpoint, action, id, params)

confirm_path(conn_or_endpoint, action)

confirm_path(conn_or_endpoint, action, params)

confirm_url(conn_or_endpoint, action)

confirm_url(conn_or_endpoint, action, params)

feed_path(conn_or_endpoint, action)

feed_path(conn_or_endpoint, action, params)

feed_path(conn_or_endpoint, action, id, params)

feed_url(conn_or_endpoint, action)

feed_url(conn_or_endpoint, action, params)

feed_url(conn_or_endpoint, action, id, params)

follower_path(conn_or_endpoint, action, id)

follower_path(conn_or_endpoint, action, id, params)

follower_url(conn_or_endpoint, action, id)

follower_url(conn_or_endpoint, action, id, params)

friend_path(conn_or_endpoint, action)

friend_path(conn_or_endpoint, action, id)

friend_path(conn_or_endpoint, action, id, params)

friend_url(conn_or_endpoint, action)

friend_url(conn_or_endpoint, action, id)

friend_url(conn_or_endpoint, action, id, params)

media_path(conn_or_endpoint, action)

media_path(conn_or_endpoint, action, params)

media_url(conn_or_endpoint, action)

media_url(conn_or_endpoint, action, params)

meta_path(conn_or_endpoint, action)

meta_path(conn_or_endpoint, action, params)

meta_url(conn_or_endpoint, action)

meta_url(conn_or_endpoint, action, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_path(conn_or_endpoint, action, path, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action, path, params)

password_reset_path(conn_or_endpoint, action)

password_reset_path(conn_or_endpoint, action, params)

password_reset_url(conn_or_endpoint, action)

password_reset_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

post_path(conn_or_endpoint, action)

post_path(conn_or_endpoint, action, id)

post_path(conn_or_endpoint, action, id, params)

post_url(conn_or_endpoint, action)

post_url(conn_or_endpoint, action, id)

post_url(conn_or_endpoint, action, id, params)

reaction_path(conn_or_endpoint, action, post_id, name)

reaction_path(conn_or_endpoint, action, post_id, name, params)

reaction_url(conn_or_endpoint, action, post_id, name)

reaction_url(conn_or_endpoint, action, post_id, name, params)

session_path(conn_or_endpoint, action)

session_path(conn_or_endpoint, action, params)

session_url(conn_or_endpoint, action)

session_url(conn_or_endpoint, action, params)

setting_path(conn_or_endpoint, action)

setting_path(conn_or_endpoint, action, params)

setting_url(conn_or_endpoint, action)

setting_url(conn_or_endpoint, action, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

tag_path(conn_or_endpoint, action)

tag_path(conn_or_endpoint, action, params)

tag_path(conn_or_endpoint, action, id, params)

tag_url(conn_or_endpoint, action)

tag_url(conn_or_endpoint, action, params)

tag_url(conn_or_endpoint, action, id, params)

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action, id)

user_path(conn_or_endpoint, action, id, params)

user_url(conn_or_endpoint, action, id)

user_url(conn_or_endpoint, action, id, params)